archive/etl_pipeline/requirements.txt
# Strategic Narrative Intelligence ETL Pipeline Dependencies
# Production requirements

# Core framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1

# Database and caching
psycopg2-binary==2.9.9
redis==5.0.1
pgvector==0.2.4

# Task queue and scheduling
celery==5.3.4
kombu==5.3.4
flower==2.0.1
crontab==1.0.1

# HTTP and networking
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# Data processing and validation
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
validators==0.22.0

# Feed parsing and web scraping
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
tldextract==5.1.1

# Language detection and processing
langdetect==1.0.9
python-magic==0.4.27

# NLP and ML dependencies
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Security and authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Async utilities
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Development and testing utilities
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Content extraction and processing
trafilatura>=1.8.0

# Additional utilities
python-json-logger==2.0.7
typing-extensions==4.8.0
email-validator==2.1.0